Remove prototypes for configuration and steps
The prototypes for the configuration and workflow steps have been
removed from the project. While experimenting with the configuration, we
learned that we have too little knowledge about both the domain and
dynamic dispatch in Rust to make this approach work. We will therefore
focus on the state machine for workflows for now, and once that is
understood reintroduce the configuration.