Open source the project (#9)
We have finished our experiments, and are ready to start working on the
real thing. This seems like a good moment to open source the project and
start working in public.

As per the Rust conventions, the project is dual licensed under the
terms of the MIT or Apache 2.0 license.