Create shared state (#10)
Individual tasks inside an automaton can share data by putting it into a
shared state. The state indexes the data based on its type, which
enables the Rust compiler to enforce its safety guarantees when reading
and writing to the state. The downside of this approach is that each
data type can only be stored once. This can be worked around with the
newtype pattern, but might still cause issues in the long run.

An alternative approach would have been to serialize the data before
storing it, and deserializing it when it is retrieved. This comes with a
performance cost, but is a very flexible solution.