Create trait for automatons (#12)
Automatons execute a series of tasks. New traits have been created for
both the automaton and the tasks to describe the behavior of each
individual component as well as the behavior of the system as a whole.

What is missing are events that trigger the execution of an automaton.
These will be addresses next. And the implementation can be improved for
better observability, which will make operation easier.