Report test coverage to Coveralls (#13)
We are switching from Codecov to Coveralls for the reporting and
monitoring of test coverage. Codecov has been very buggy lately, and
most recently has simply shown a test coverage of 100% in most places.
When digging into it, the correct coverage of ~80% can be found, but
the user experience has been pretty bad.