Create macros to generate identifier types (#17)
GitHub assigns unique numerical ids to most if not all resources on its
platform. And user-facing resources often have a human-readable
identifier as well, for example a name.

These identifiers could be represented by primitivate types, for example
integers and strings. But the Rust compiler wouldn't be able to ensure
that they are used correctly. Eventually, this would cause bugs and
potentially even security vulnerabilities.

Instead, we are leveraging the Rust type system. For each identifier, a
newtype is generated that represents it. For example, the user id might
be represented by a `UserId` type, while a `RepositoryId` is used with
repositories. This ensures that identifiers are not accidentally mixed.

The implementation for these identifiers is always the same, so a macro
rule has been created that can generate these identifiers. And a very
similar rule has been added for textual identifiers, e.g. names.