Rename libraries
The libraries have been renamed due to a naming conflict on crates.io.
An `automaton` crate [^1] already exists, and while it looked like a
placeholder at first, we discovered that it is actually used by a
handful of other projects [^2]. Our intention was to ask the owner to
transfer ownership of the crate, but given that it had some usage at
some point in time, this feels wrong to do.

We will therefore adjust the name (slightly) to avoid the conflict.

[^1]: https://crates.io/crates/automaton
[^2]: https://github.com/carllerche/automaton/network/dependents