Create resource for check runs (#33)
A check run is an individual test that is part of a check suite. Each
run includes a status and conclusion, and can optionally include
additional information in an output field.

The output field is always included in check run events, but set to
empty (i.e. `null`) values when the user has not provided any details. A
custom deserializer has been implemented that can handle this and turn
the output into an `Option`. This cleans up the API and avoids turning
every field of the `CheckRunOutput` into an option.