Automatically deploy default branch to staging
We plan to run the automatons framework in (at least) two different
environments: staging and production. The staging environment will run
the latest chances that have been merged into the default branch, while
production will be periodically updated with a tested and stable
version.

A new GitHub Action has been set up that builds the ingress binary for
GitHub, and deploys it to an AWS Lambda function. We consider this a
proof of concept, and plan to eventually refactor the deployment to use
something like AWS CodePipeline.