Simplify task initialization (#41)
The initialization of tasks has been simplified by removing the `init`
method. The initial ideas was that tasks can be initialized from the
state, but in practice this caused issues when a task wanted to keep
references to data in the state. We are thus falling back to a simpler
implementation that reduces the `Task` trait to the `execute` method.
Tasks can initialize the data they need in that method, which removes
any issues with lifetimes.