Re-implement ingress as an HTTP webserver
Originally, we intended to run the ingress component of the AWS runtime
as a serverless AWS Lambda function. This would have provided almost
endless scalability, which is a great feature for an ingress gateway.

As it turns out, though, writing a Lambda function in Rust is not
documented well. And we are still learning about the intricacies of AWS
and how to deploy applications there. For both reasons, the decision has
been made to postpone the AWS Lambda function and go with a simple HTTP
server that we can run as a Docker container.