Remove runtime implementation (#49)
The initial work on a runtime for automatons on AWS has been removed
from the repository. The implementation is highly customized to our own
needs at this time, and not suited for self-hosting or external
contributions. Once the implementation has stabilized and we understand
the features that the runtime needs and provides, we will reconsider
open sourcing it in parts or full.