Redesign internal state machine in automatons (#55)
After some consideration, the decision has been made to refactor the
state machine in the `automatons` crate.

The previous design was optimized for loose coupling of tasks, but it
came with considerable downsides. First, automatons were only able to
execute a sequential series of tasks. This works for simple automations,
but is not flexible enough to implement more complex workflows. And
second, the state was abused as a means to pass data between tasks. The
design didn't leverage the strength of the Rust compiler to ensure that
each task had the data that it needed at the time that it was run.

We are taking a step back and going back to an earlier design that was
used in `devxbots/octox` [^1]. Fundamentally, the GitHub integration
will only implement abstractions around GitHub's API. Implementing the
state machine with its states will be the responsibility of the
individual bots. The downside of this approach is that states are merely
thin wrappers around the GitHub actions, but it does give bots a lot
more flexibility in how to implement their states.

[^1]: https://github.com/devxbots/octox